more c style casts (#322)

+++ chakracore-cxx/lib/Common/Common/SmartFpuControl.h
@@ -11,7 +11,7 @@
 template <bool enabled>
 class SmartFPUControlT
 {
-    static const uint INVALID_FPUCONTROL = (uint)-1;
+    static const uint INVALID_FPUCONTROL = -1u;
 
 public:
